Refactoring sync integ tests
