+++ tests/integration/sync/test_sync_infra.py
@@ -5,25 +5,17 @@ import logging
 import json
 import shutil
 import tempfile
-import time
-import uuid
 from pathlib import Path
 from unittest import skipIf
 
-import boto3
-from botocore.exceptions import ClientError
-from botocore.config import Config
 from parameterized import parameterized
 
-from samcli.lib.bootstrap.bootstrap import SAM_CLI_STACK_NAME
 from samcli.lib.utils.resources import (
     AWS_APIGATEWAY_RESTAPI,
     AWS_LAMBDA_FUNCTION,
     AWS_STEPFUNCTIONS_STATEMACHINE,
 )
-from tests.integration.buildcmd.build_integ_base import BuildIntegBase
 from tests.integration.sync.sync_integ_base import SyncIntegBase
-from tests.integration.package.package_integ_base import PackageIntegBase
 
 from tests.testing_utils import RUNNING_ON_CI, RUNNING_TEST_FOR_MASTER_ON_CI, RUN_BY_CANARY
 from tests.testing_utils import run_command_with_input
@@ -32,7 +24,7 @@ from tests.testing_utils import run_command_with_input
 # This is to restrict package tests to run outside of CI/CD, when the branch is not master or tests are not run by Canary
 SKIP_SYNC_TESTS = RUNNING_ON_CI and RUNNING_TEST_FOR_MASTER_ON_CI and not RUN_BY_CANARY
 IS_WINDOWS = platform.system().lower() == "windows"
-CFN_SLEEP = 3
+# Some wait time for code updates to be reflected on each service
 LAMBDA_SLEEP = 3
 API_SLEEP = 5
 SFN_SLEEP = 5
@@ -42,40 +34,7 @@ LOG = logging.getLogger(__name__)
 
 
 @skipIf(SKIP_SYNC_TESTS, "Skip sync tests in CI/CD only")
-class TestSync(BuildIntegBase, PackageIntegBase, SyncIntegBase):
-    @classmethod
-    def setUpClass(cls):
-        PackageIntegBase.setUpClass()
-
-        cls.test_data_path = Path(__file__).resolve().parents[1].joinpath("testdata", "sync")
-
-    def setUp(self):
-        self.cfn_client = boto3.client("cloudformation")
-        self.ecr_client = boto3.client("ecr")
-        self.lambda_client = boto3.client("lambda")
-        self.api_client = boto3.client("apigateway")
-        self.sfn_client = boto3.client("stepfunctions")
-        self.sns_arn = os.environ.get("AWS_SNS")
-        self.stacks = []
-        time.sleep(CFN_SLEEP)
-        self.s3_prefix = uuid.uuid4().hex
-        super().setUp()
-
-    def tearDown(self):
-        shutil.rmtree(os.path.join(os.getcwd(), ".aws-sam", "build"), ignore_errors=True)
-        for stack in self.stacks:
-            # because of the termination protection, do not delete aws-sam-cli-managed-default stack
-            stack_name = stack["name"]
-            if stack_name != SAM_CLI_STACK_NAME:
-                region = stack.get("region")
-                cfn_client = (
-                    self.cfn_client if not region else boto3.client("cloudformation", config=Config(region_name=region))
-                )
-                ecr_client = self.ecr_client if not region else boto3.client("ecr", config=Config(region_name=region))
-                self._delete_companion_stack(cfn_client, ecr_client, self._stack_name_to_companion_stack(stack_name))
-                cfn_client.delete_stack(StackName=stack_name)
-        super().tearDown()
-
+class TestSyncInfra(SyncIntegBase):
     @skipIf(
         IS_WINDOWS,
         "Skip sync ruby tests in windows",
@@ -238,163 +197,6 @@ Requires capabilities : [CAPABILITY_AUTO_EXPAND]",
             str(sync_process_execute.stderr),
         )
 
-    def test_sync_code_base(self):
-        with tempfile.TemporaryDirectory() as temp:
-            temp_dir = Path(temp).joinpath("code")
-            shutil.copytree(self.test_data_path.joinpath("code").joinpath("before"), temp_dir)
-
-            template = "template-python.yaml"
-            template_path = temp_dir.joinpath(template)
-            stack_name = self._method_to_stack_name(self.id())
-            self.stacks.append({"name": stack_name})
-
-            # Run infra sync
-            sync_command_list = self.get_sync_command_list(
-                template_file=template_path,
-                code=False,
-                watch=False,
-                base_dir=temp_dir,
-                dependency_layer=True,
-                stack_name=stack_name,
-                parameter_overrides="Parameter=Clarity",
-                image_repository=self.ecr_repo_name,
-                s3_prefix=self.s3_prefix,
-                kms_key_id=self.kms_key,
-                tags="integ=true clarity=yes foo_bar=baz",
-            )
-
-            sync_process_execute = run_command_with_input(sync_command_list, "y\n".encode())
-            self.assertEqual(sync_process_execute.process.returncode, 0)
-            self.assertIn("Stack creation succeeded. Sync infra completed.", str(sync_process_execute.stderr))
-
-            shutil.rmtree(temp_dir.joinpath("function"), ignore_errors=True)
-            shutil.copytree(
-                self.test_data_path.joinpath("code").joinpath("after").joinpath("function"),
-                temp_dir.joinpath("function"),
-            )
-            self._test_sync_code_function(template_path, stack_name)
-
-            shutil.rmtree(temp_dir.joinpath("layer"), ignore_errors=True)
-            shutil.copytree(
-                self.test_data_path.joinpath("code").joinpath("after").joinpath("layer"), temp_dir.joinpath("layer")
-            )
-            self._test_sync_code_layer(template_path, stack_name)
-
-            shutil.rmtree(temp_dir.joinpath("apigateway"), ignore_errors=True)
-            shutil.copytree(
-                self.test_data_path.joinpath("code").joinpath("after").joinpath("apigateway"),
-                temp_dir.joinpath("apigateway"),
-            )
-            self._test_sync_code_rest_api(template_path, stack_name, temp_dir)
-
-            shutil.rmtree(temp_dir.joinpath("statemachine"), ignore_errors=True)
-            shutil.copytree(
-                self.test_data_path.joinpath("code").joinpath("after").joinpath("statemachine"),
-                temp_dir.joinpath("statemachine"),
-            )
-            self._test_sync_code_state_machine(template_path, stack_name, temp_dir)
-
-    def _test_sync_code_function(self, template_path, stack_name):
-        # Run code sync
-        sync_command_list = self.get_sync_command_list(
-            template_file=template_path,
-            code=True,
-            watch=False,
-            resource="AWS::Serverless::Function",
-            dependency_layer=True,
-            stack_name=stack_name,
-            parameter_overrides="Parameter=Clarity",
-            image_repository=self.ecr_repo_name,
-            s3_prefix=self.s3_prefix,
-            kms_key_id=self.kms_key,
-            tags="integ=true clarity=yes foo_bar=baz",
-        )
-        sync_process_execute = run_command_with_input(sync_command_list, "y\n".encode())
-        self.assertEqual(sync_process_execute.process.returncode, 0)
-
-        # CFN Api call here to collect all the stack resources
-        self.stack_resources = self._get_stacks(stack_name)
-        # Lambda Api call here, which tests both the python function and the layer
-        lambda_functions = self.stack_resources.get(AWS_LAMBDA_FUNCTION)
-        for lambda_function in lambda_functions:
-            lambda_response = json.loads(self._get_lambda_response(lambda_function))
-            self.assertIn("extra_message", lambda_response)
-            self.assertEqual(lambda_response.get("message"), "8")
-
-    def _test_sync_code_layer(self, template_path, stack_name):
-        # Run code sync
-        sync_command_list = self.get_sync_command_list(
-            template_file=template_path,
-            code=True,
-            watch=False,
-            resource="AWS::Serverless::LayerVersion",
-            dependency_layer=True,
-            stack_name=stack_name,
-            parameter_overrides="Parameter=Clarity",
-            image_repository=self.ecr_repo_name,
-            s3_prefix=self.s3_prefix,
-            kms_key_id=self.kms_key,
-            tags="integ=true clarity=yes foo_bar=baz",
-        )
-        sync_process_execute = run_command_with_input(sync_command_list, "y\n".encode())
-        self.assertEqual(sync_process_execute.process.returncode, 0)
-
-        # CFN Api call here to collect all the stack resources
-        self.stack_resources = self._get_stacks(stack_name)
-        # Lambda Api call here, which tests both the python function and the layer
-        lambda_functions = self.stack_resources.get(AWS_LAMBDA_FUNCTION)
-        time.sleep(LAMBDA_SLEEP)
-        for lambda_function in lambda_functions:
-            lambda_response = json.loads(self._get_lambda_response(lambda_function))
-            self.assertIn("extra_message", lambda_response)
-            self.assertEqual(lambda_response.get("message"), "9")
-
-    def _test_sync_code_rest_api(self, template_path, stack_name, base_dir):
-        # Run code sync
-        sync_command_list = self.get_sync_command_list(
-            template_file=template_path,
-            code=True,
-            watch=False,
-            base_dir=base_dir,
-            resource="AWS::Serverless::Api",
-            stack_name=stack_name,
-            parameter_overrides="Parameter=Clarity",
-            image_repository=self.ecr_repo_name,
-            s3_prefix=self.s3_prefix,
-            kms_key_id=self.kms_key,
-            tags="integ=true clarity=yes foo_bar=baz",
-        )
-        sync_process_execute = run_command_with_input(sync_command_list, "y\n".encode())
-        self.assertEqual(sync_process_execute.process.returncode, 0)
-
-        time.sleep(API_SLEEP)
-        # ApiGateway Api call here, which tests the RestApi
-        rest_api = self.stack_resources.get(AWS_APIGATEWAY_RESTAPI)[0]
-        self.assertEqual(self._get_api_message(rest_api), '{"message": "hello!!!"}')
-
-    def _test_sync_code_state_machine(self, template_path, stack_name, base_dir):
-        # Run code sync
-        sync_command_list = self.get_sync_command_list(
-            template_file=template_path,
-            code=True,
-            watch=False,
-            base_dir=base_dir,
-            resource="AWS::Serverless::StateMachine",
-            stack_name=stack_name,
-            parameter_overrides="Parameter=Clarity",
-            image_repository=self.ecr_repo_name,
-            s3_prefix=self.s3_prefix,
-            kms_key_id=self.kms_key,
-            tags="integ=true clarity=yes foo_bar=baz",
-        )
-        sync_process_execute = run_command_with_input(sync_command_list, "y\n".encode())
-        self.assertEqual(sync_process_execute.process.returncode, 0)
-
-        # SFN Api call here, which tests the StateMachine
-        time.sleep(SFN_SLEEP)
-        state_machine = self.stack_resources.get(AWS_STEPFUNCTIONS_STATEMACHINE)[0]
-        self.assertEqual(self._get_sfn_response(state_machine), '"World has been updated!!"')
-
     @parameterized.expand(
         [
             (
@@ -512,94 +314,3 @@ Requires capabilities : [CAPABILITY_AUTO_EXPAND]",
                 lambda_response = json.loads(self._get_lambda_response(lambda_function))
                 self.assertIn("extra_message", lambda_response)
                 self.assertEqual(lambda_response.get("message"), "9")
-
-    def _get_stacks(self, stack_name):
-        try:
-            physical_ids = {}
-            response = self.cfn_client.describe_stack_resources(StackName=stack_name).get("StackResources", {})
-            for resource in response:
-                resource_type = resource.get("ResourceType")
-                if resource_type == "AWS::CloudFormation::Stack":
-                    nested_stack_physical_id = resource.get("PhysicalResourceId")
-                    nested_stack_name = nested_stack_physical_id.split("/")[1]
-                    nested_stack_physical_ids = self._get_stacks(nested_stack_name)
-                    for nested_resource_type, nested_physical_ids in nested_stack_physical_ids.items():
-                        if nested_resource_type not in physical_ids:
-                            physical_ids[nested_resource_type] = []
-                        physical_ids[nested_resource_type] += nested_physical_ids
-                    continue
-                if resource_type not in physical_ids:
-                    physical_ids[resource.get("ResourceType")] = []
-                physical_ids[resource_type].append(resource.get("PhysicalResourceId"))
-            return physical_ids
-        except ClientError:
-            pass
-        return None
-
-    def _get_lambda_response(self, lambda_function):
-        try:
-            lambda_response = self.lambda_client.invoke(FunctionName=lambda_function, InvocationType="RequestResponse")
-            payload = json.loads(lambda_response.get("Payload").read().decode("utf-8"))
-            return payload.get("body")
-        except ClientError as ce:
-            LOG.error(ce)
-        except Exception as e:
-            LOG.error(e)
-        return ""
-
-    def _get_api_message(self, rest_api):
-        try:
-            api_resource = self.api_client.get_resources(restApiId=rest_api)
-            for item in api_resource.get("items"):
-                if "GET" in item.get("resourceMethods", {}):
-                    resource_id = item.get("id")
-                    break
-            self.api_client.flush_stage_cache(restApiId=rest_api, stageName="prod")
-        except ClientError as ce:
-            LOG.error(ce)
-        except Exception as e:
-            LOG.error(e)
-
-        # RestApi deployment needs a wait time before being responsive
-        count = 0
-        while count < 20:
-            try:
-                time.sleep(1)
-                api_response = self.api_client.test_invoke_method(
-                    restApiId=rest_api, resourceId=resource_id, httpMethod="GET"
-                )
-                return api_response.get("body")
-            except ClientError as ce:
-                if count == 20:
-                    LOG.error(ce)
-                # This test is very unstable, any fixed wait time cannot guarantee a successful invocation
-                if "Invalid Method identifier specified" in ce.response.get("Error", {}).get("Message", ""):
-                    if count == 20:
-                        LOG.error(
-                            "The deployed changes are not callable on the client yet, skipping the RestApi invocation"
-                        )
-                        return '{"message": "hello!!"}'
-            except Exception as e:
-                if count == 20:
-                    LOG.error(e)
-            count += 1
-        return ""
-
-    def _get_sfn_response(self, state_machine):
-        try:
-            timestamp = str(int(time.time() * 1000))
-            name = f"sam_integ_test_{timestamp}"
-            sfn_execution = self.sfn_client.start_execution(stateMachineArn=state_machine, name=name)
-            execution_arn = sfn_execution.get("executionArn")
-            count = 0
-            while count < 20:
-                time.sleep(1)
-                execution_detail = self.sfn_client.describe_execution(executionArn=execution_arn)
-                if execution_detail.get("status") == "SUCCEEDED":
-                    return execution_detail.get("output")
-                count += 1
-        except ClientError as ce:
-            LOG.error(ce)
-        except Exception as e:
-            LOG.error(e)
-        return ""
